Rename debug_diagram to PipelineDiagram and move into pipeline package
