+++ v2/tools/generator/internal/codegen/code_generator.go
@@ -231,8 +231,8 @@ func (generator *CodeGenerator) Generate(ctx context.Context) error {
 	if generator.debugReporter != nil {
 		// Generate a diagram containing our stages
 		outputFolder := generator.debugReporter.outputFolder
-		diagram := newDebugDiagram(outputFolder)
-		err := diagram.writeDiagram(generator.pipeline)
+		diagram := pipeline.NewPipelineDiagram(outputFolder)
+		err := diagram.WriteDiagram(generator.pipeline)
 		if err != nil {
 			return errors.Wrapf(err, "failed to generate diagram")
 		}
